backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0.post1
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-multipart==0.0.6
gunicorn==21.2.0
python-jose[cryptography]==3.3.0
pytest-cov==4.1.0
alembic==1.13.1
black==23.11.0
flake8==6.1.0
